Update TODO file. Clean up and reorganize google module. Move image classes to images module. Write tests for future currency module.
